Add files via upload
As the objects pass through the conveyor belt, they are captured through the camera. The task is to differentiate among the different objects and separate the object that is recognized as a square.
Through this code, the pipelines are created to first detect an object and classify them based on their shapes; the object detection module of the TensorFlow library is used for this purpose.