Add files via upload
This video is a Unity 3D implementation of the Digital Twin Model. The video shows the speed controls of the conveyor belt (ON, OFF, HIGH, LOW) and the working of the robotic arm - when it detects a square object, the end-effector of the robotic arm separates that object from the other shaped objects.