Add files via upload
This video is the physical implementation of the Digital Twin Model.
The objects (star, square and circle) are placed on the conveyor belt, the star is collected on the other end. When the square shaped object is detected by the camera on the robotic arm, the end effector pushes it away from the conveyor belt and it will be separated from the rest of the objects. The circle shaped object passes through the conveyor belt and reaches the other end as star shaped.